Get block ids, descriptions, and elements

Retrieve each block's id, as generated by Qualtrics, and description, as assigned by the user in the control panel. Optionally, if argument elements = TRUE, blocks also shows the elements of each block (questions and page breaks), in a long table of block-element pairs.

blocks(design_object, elements = FALSE)

Arguments

design_object

A qualtrics_design-class object retrieved from Qualtrics by design.

elements

Show the elements of each block, or don't (default).

Value

A table giving the block_id and block_description of each block, and optionally its elements.