Drop columns from survey responses

drop_sensitive drops the columns that give approximate location, IP address, contact name, and contact email from a table of survey responses. drop_meta also drops other standard columns that don't represent question responses. keep_questions is most strict, keeping only the columns for survey questions; it drops columns with embedded data, for example.

drop_sensitive(tbl)

drop_meta(tbl)

keep_questions(tbl, design_object)

Arguments

tbl
A data.frame of survey responses.

design_object
A qualtrics_design-class object retrieved from Qualtrics by design.

Value

The data.frame without the dropped columns.

Details

The following metadata columns are dropped by drop_sensitive, if they exist in the table:

LocationLatitude

LocationLongitude

LocationAccuracy

IPAddress

RecipientLastName

RecipientFirstName

RecipientEmail

drop_meta drops those columns and any of the following:

ExternalDataReference

EndDate

Finished

ResponseID

ResponseSet

StartDate

Status