The qsurvey package: overview

qsurvey is a toolkit for working with the Qualtrics survey platform and its data in R. It focuses on testing and review of surveys before fielding, and analysis of responses afterward.

Details

Examine branching, randomization, and question attributes

Represent survey flows as directed graphs for plotting or interactive review in Shiny

Download and manipulate response data

For more see the http://jdunham.io/qsurvey.

A https://www.qualtrics.com/support/integrations/api-integration/api-integration is needed to communicate with the survey platform. Set the environment variable QUALTRICS_KEY to your key value. You can do this during R startup (recommended), interactively with Sys.setenv, or through key_from_file.