Send a request to the Qualtrics API

These are lower-level functions for direct access to API actions. They may be useful in programming or when higher-level package functions are not available. qget and qpost are wrappers for request.

request(verb = "GET", action = NULL, key = Sys.getenv("QUALTRICS_KEY"),
  subdomain = Sys.getenv("QUALTRICS_SUBDOMAIN"), api_url = NULL,
  verbose = FALSE, ...)

qget(action = NULL, as = "parsed", key = Sys.getenv("QUALTRICS_KEY"),
  subdomain = Sys.getenv("QUALTRICS_SUBDOMAIN"), ...)

qpost(action = NULL, as = "parsed", key = Sys.getenv("QUALTRICS_KEY"),
  subdomain = Sys.getenv("QUALTRICS_SUBDOMAIN"), ...)

Arguments

verb

Name of verb to use.

action

An API action (like "surveys").

key

A Qualtrics API key (by default, the value of the environment variable QUALTRICS_KEY).

subdomain

A Qualtrics subdomain (by default, the value of the environment variable QUALTRICS_SUBDOMAIN).

api_url

A complete API URL. If used, argument action and subdomain are ignored.

verbose

Output API calls to stderr.

...

Further arguments to GET or POST, depending on verb.

as

Desired type of content. See content.

Value

For request, a response object. For qget and qpost, its content as extracted by content.

Details

For help with subdomains, see the https://api.qualtrics.com/docs/root-url. Each Qualtrics account is assigned a subdomain, and using another will work but produce a warning.