Download survey responses

Retrieve the responses to a survey.

responses(id, format = "json", use_labels = TRUE, verbose = TRUE,
  key = Sys.getenv("QUALTRICS_KEY"), ...)

Arguments

id

A Qualtrics survey identifier.

use_labels

Use question labels and choice descriptions (default), instead of question and identifiers.

verbose

Print progress.

key

A Qualtrics API key (by default, the value of the environment variable QUALTRICS_KEY).

...

Additional parameters for the responseexports API.

Value

A data.table of survey responses.

Details

responses requests that Qualtrics prepare a zipfile for download; once it is ready, the archive is downloaded to a location given by tempfile and unzipped; its JSON contents read and parsed; and the temporary files deleted.

By default, descriptive question labels and choice descriptions (e.g., favorability and strongly approve) are requested instead of question ids and choice codes (e.g., QID1 and 5). This is controlled by argument use_labels.

Other parameters that affect the format of the survey responses are available. For example, passing seenUnansweredRecode = "99" in the call to responses will show "99" as the response to unanswered questions instead of NA. See the https://api.qualtrics.com/docs/json for the details of these parameters. The parameters that cannot be changed via ... are format, surveyId, and useLabels, the last of which is exposed as argument use_labels.

See also

Retrieve a survey's questions or question choices.