Download survey metadata

Retrieve names, identifiers, and other metadata for all available surveys. The result is sorted by the last_modified timestamp in descending order.

surveys(key = Sys.getenv("QUALTRICS_KEY"),
  subdomain = Sys.getenv("QUALTRICS_SUBDOMAIN"))

Arguments

key

A Qualtrics API key (by default, the value of the environment variable QUALTRICS_KEY).

subdomain

A Qualtrics subdomain (by default, the value of the environment variable QUALTRICS_SUBDOMAIN).

Value

A data.table of survey metadata

Examples

surveys()

#>                    id                 name           owner_id
#> 1: SV_0CGgkDZJaUvxnGl     Student Feedback UR_8J1114L8aAeMCPP
#> 2: SV_cIKqX1B5wDO3wKV    Course Evaluation UR_8J1114L8aAeMCPP
#> 3: SV_5goFlnLiTZWGI1D Patient Satisfaction UR_8J1114L8aAeMCPP
#> 4: SV_eWLnQvYuRfIKQRL   Country & US State UR_8J1114L8aAeMCPP
#> 5: SV_0VVlb9QwJ4bsBKZ         Demographics UR_8J1114L8aAeMCPP
#>          last_modified is_active
#> 1: 2016-11-24 03:36:38     FALSE
#> 2: 2016-11-24 03:36:29     FALSE
#> 3: 2016-11-24 03:36:24     FALSE
#> 4: 2016-11-24 03:36:19     FALSE
#> 5: 2016-11-24 03:36:11     FALSE